requirements.txt
grpcio~=1.70.0
grpcio-tools~=1.70.0
netifaces~=0.11.0
protobuf~=5.29.3
pydantic~=2.10.6
PyQt5~=5.15.11
pytest~=8.3.4
pyyaml~=6.0.2
